Sample spatial field

Create a sample of a spatial field.

sam_field(x, size, type, value, nclusters)

Arguments

x

Any object that can be coerced to an sf object.

size

Number of samples to create.

type

Expected spatial distribution of the samples. Possible values are: "jittered", "random", "clust".

value

Amount of jitter to apply to the samples (in map units; only used when type = "jittered") or radius of the buffer around each cluster (only used when type = "clust").

nclusters

Number of clusters to simulate. Only used when type = "clust".

Examples

rast_grid = terra::rast(ncols = 300, nrows = 100, xmin = 0, xmax = 300, ymin = 0, ymax = 100)
sam_field(rast_grid, 100, "jittered", 5)
#> Simple feature collection with 85 features and 0 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 6.371558 ymin: 9.430832 xmax: 283.3544 ymax: 88.38453
#> CRS:           NA
#> First 10 features:
#>                     geometry
#> 1    POINT (7.912142 11.283)
#> 2  POINT (20.79713 15.86894)
#> 3  POINT (36.61954 9.737413)
#> 4  POINT (58.52999 19.12196)
#> 5  POINT (76.16434 10.64163)
#> 6  POINT (91.40475 14.33684)
#> 7  POINT (113.1564 17.61247)
#> 8  POINT (130.8733 16.33397)
#> 9  POINT (149.2146 11.56441)
#> 10 POINT (159.5385 14.62171)
vect_grid = sf::st_as_sf(terra::as.polygons(rast_grid))
sam_field(vect_grid, 100, "random")
#> Simple feature collection with 100 features and 0 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 1.348892 ymin: 1.476902 xmax: 299.8957 ymax: 98.65087
#> CRS:           NA
#> First 10 features:
#>                     geometry
#> 1  POINT (1.348892 19.62195)
#> 2  POINT (283.0549 96.86375)
#> 3  POINT (131.4412 38.70963)
#> 4   POINT (225.181 65.03439)
#> 5  POINT (200.3447 81.45962)
#> 6   POINT (122.392 7.096477)
#> 7  POINT (105.3746 52.68303)
#> 8  POINT (221.4275 76.34748)
#> 9  POINT (199.2856 43.53866)
#> 10 POINT (25.56741 55.24723)